Got face recogition working on vue
