+++ package.json
@@ -11,6 +11,7 @@
     "bootstrap": "^5.1.3",
     "bootstrap-vue": "^2.21.2",
     "core-js": "^3.8.3",
+    "face-api.js": "^0.22.2",
     "vue": "^2.6.14",
     "vue-router": "^3.5.1",
     "vue-web-cam": "^1.9.0",
