+++ public/assets/labeled_images/Black Widow/2.jpg
