+++ public/assets/labeled_images/Black Widow/5.jpg
