+++ public/assets/labeled_images/Captain America/1.jpg
