+++ public/assets/labeled_images/Captain America/2.jpg
