+++ public/assets/labeled_images/Captain America/4.jpg
