+++ public/assets/labeled_images/Captain America/5.jpg
