+++ public/assets/labeled_images/Thor/1.jpg
