+++ public/assets/labeled_images/Thor/2.jpg
