+++ public/assets/labeled_images/Thor/3.jpg
