+++ public/assets/labeled_images/Thor/4.jpg
