+++ public/assets/labeled_images/Tommy/1.jpg
