+++ public/assets/labeled_images/Tony Stark/1.jpg
