+++ public/assets/labeled_images/Tony Stark/2.jpg
