+++ public/assets/models/ssd_mobilenetv1_model-shard2
