+++ public/index.html
@@ -11,14 +11,14 @@
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css"
     integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
   <title><%= htmlWebpackPlugin.options.title %></title>
+  <script defer src="./assets/face-api.min.js"></script>
+  <script defer src="./assets/script.js"></script>
 </head>
 
 <style>
   body {
     padding-top: 100px
   }
-
-  
 </style>
 
 <body>
